+++ packages/react/src/connectToStores.ts
@@ -1,4 +1,4 @@
-import { StoreClass, StoreGetter } from '@fleur/fleur'
+import { StoreGetter } from '@fleur/fleur'
 import * as React from 'react'
 
 import { useStore } from './useStore'
@@ -11,15 +11,12 @@ type ConnectedComponent<Props, MappedProps> = React.ComponentType<
 >
 
 export const connectToStores = <Props, MappedProps = {}>(
-  stores: StoreClass[],
   mapStoresToProps: StoreToPropMapper<Props, MappedProps>,
 ) => <ComponentProps extends object>(
   Component: React.ComponentType<ComponentProps>,
 ): ConnectedComponent<ComponentProps, MappedProps> => {
   return React.forwardRef((props: any, ref) => {
-    const mappedProps = useStore(stores, getStore =>
-      mapStoresToProps(getStore, props),
-    )
+    const mappedProps = useStore(getStore => mapStoresToProps(getStore, props))
 
     return React.createElement(Component, { ref, ...props, ...mappedProps })
   })