+++ packages/react/src/ssr.spec.tsx
@@ -1,3 +1,5 @@
+/** @jest-environment node */
+
 import Fleur, { action, listen, operation, Store } from '@fleur/fleur'
 import * as cheerio from 'cheerio'
 import * as express from 'express'
@@ -22,7 +24,7 @@ describe('Sever side rendering', () => {
 
     class TestStore extends Store {
       public static storeName = 'TestStore'
-      protected state = { count: 0 }
+      public state = { count: 0 }
       private increase = listen(increaseIdent, ({ increase }) => {
         this.updateWith(d => (d.count += increase))
       })
@@ -32,7 +34,7 @@ describe('Sever side rendering', () => {
     }
 
     const Component = () => {
-      const { count } = useStore([TestStore], getStore => ({
+      const { count } = useStore(getStore => ({
         count: getStore(TestStore).getCount(),
       }))