+++ packages/create-fleur-next-app/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.3.0
+
+- [#160](https://github.com/ra-gg/fleur/pull/160) Bump up `@fleur/react` version to 3.0.0 and follow API changes
+
 ## 0.2.0
 
 - Update `@fleur/fleur-react` to `@fleur/react`