+++ packages/route-store-dom/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2.2.0
+
+- [#160](https://github.com/ra-gg/fleur/pull/160) Upgrade minimum required `@fleur/react` version to 3.0.0 and follow API changes
+
 ### 2.1.1
 
 - Update deps